pyproject.toml
[project]
name = "froggy"
version = "0.0.1"
description = "Chat with your local models"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "huggingface-hub>=0.25.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
gpu = [
    "torch",
    "transformers",
    "accelerate",
    "peft",
    "huggingface-hub",
]
mlx = [
    "mlx>=0.16.0",
    "mlx-lm>=0.16.0",
]
tools = [
    "duckduckgo_search>=6.0",
]

[project.scripts]
froggy = "froggy.cli:cli"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
